\name{rCodec}
\alias{rCodec}
\docType{package}
\title{rCodec: Interface to Apache Commons Codec} 
\description{
   The functions in this package allow one to use the functionality of
   the Apache Commons Codec Java library in R programs.
}
\seealso{
Functionality implemented so far:\cr
  \code{\link{codec.doubleMetaphone}}\cr
}
