\name{rCodec}
\alias{rCodec}
\docType{package}
\title{rCodec: Interface to Apache Commons Codec} 
\description{
   The functions in this package allow one to use the functionality of
   the Apache Commons Codec Java library in R programs.
}
\seealso{
Functionality implemented so far:\cr\cr
  \code{\link{codec.metaphone}}\cr
  \code{\link{codec.doubleMetaphone}}\cr
  \code{\link{codec.caverphone}}\cr
  \code{\link{codec.colognePhonetic}}\cr
  \code{\link{codec.nysiis}}\cr
  \code{\link{codec.refined.soundex}}\cr
  \code{\link{codec.refined.soundex.difference}}\cr
  \code{\link{codec.soundex}}\cr
  \code{\link{codec.soundex.difference}}\cr
}
